split todo app components
