+++ todo-app/src/app/components/TodoItem.tsx
@@ -1,0 +1,37 @@
+'use client';
+
+interface Todo {
+  _id: string;
+  title: string;
+  completed: boolean;
+}
+
+interface TodoItemProps {
+  todo: Todo;
+  onToggle: (todo: Todo) => void;
+}
+
+export default function TodoItem({ todo, onToggle }: TodoItemProps) {
+  return (
+    <li className="flex items-center justify-between p-3 bg-gray-50 dark:bg-gray-700 rounded-lg hover:bg-gray-100 dark:hover:bg-gray-600 transition-colors">
+      <div className="flex items-center gap-3">
+        <input
+          type="checkbox"
+          checked={todo.completed}
+          onChange={() => onToggle(todo)}
+          className="w-5 h-5 rounded border-gray-300 dark:border-gray-600 text-blue-600 focus:ring-blue-500"
+        />
+        <span className={`${todo.completed ? 'line-through text-gray-500 dark:text-gray-400' : 'text-gray-700 dark:text-gray-200'}`}>
+          {todo.title}
+        </span>
+      </div>
+      <span className={`px-2 py-1 text-xs rounded-full ${
+        todo.completed 
+          ? 'bg-green-100 text-green-800 dark:bg-green-800 dark:text-green-100'
+          : 'bg-yellow-100 text-yellow-800 dark:bg-yellow-800 dark:text-yellow-100'
+      }`}>
+        {todo.completed ? 'Completed' : 'Pending'}
+      </span>
+    </li>
+  );
+}
